Add hash function to allow string fnc names
Hash function was added to allow JSON with string method names
instead of function codes.

String names are human readable and easier to debug.

We expect that electrolink-server (device side) will be implemented
mostly in Lua (or maybe uPython or JerryScript) and in that case
havin string method name makes no problem (as in C where we need hash
fnc).

Signed-off-by: Drasko DRASKOVIC <drasko.draskovic@gmail.com>